[#384]; refactor: auth/login을 auth/authentication으로 병합
- LoginService가 OAuth2Service.login()을 감싸서 회원 여부만 체크하는 로직이라 authentication과 개념적으로 같음, 파일 2개짜리 하위도메인을 따로 둘 필요 없어 흡수
- 병합 과정에서 같은 패키지가 되며 불필요해진 import, 남은 빈 줄 정리
