+++ src/main/kotlin/com/yourssu/scouter/auth/user/storage/UserEntity.kt
@@ -14,7 +14,6 @@ import jakarta.persistence.Id
 import jakarta.persistence.Table
 import java.time.Instant
 
-
 @Entity
 @Table(name = "users")
 class UserEntity(
