+++ src/test/kotlin/com/yourssu/scouter/auth/authentication/business/LoginServiceTest.kt
@@ -1,7 +1,6 @@
-package com.yourssu.scouter.auth.login.business
+package com.yourssu.scouter.auth.authentication.business
 
 import com.yourssu.scouter.auth.authentication.business.dto.LoginResult
-import com.yourssu.scouter.auth.authentication.business.OAuth2Service
 import com.yourssu.scouter.auth.authentication.implement.OAuth2Type
 import com.yourssu.scouter.member.core.fixture.MemberFixtureBuilder
 import com.yourssu.scouter.member.core.implement.MemberReader
